api/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib
python-jose
python-multipart
